Add 'guest' login support. Prepare for v1.1.
An external login service may be configured to have a 'guest' user,
which is used in case when the email address from the service does
not correspond to a specific WordPress user. Guest user must not have
certain capabilities (e.g. 'edit_posts'), and they cannot access the
dashboard or update their own profile.